Updated miss placed files that occurred during first upload
