+++ src/Contracts/Resources/ResponsesContract.php
